# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API",
        "description": "Documentació oficial de l'API d'Inventari - Projecte Final",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/providers": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Obtenir llista de proveïdors",
                "operationId": "d99e139f235d5222c74789c37512fc95",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar per nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrar per estat actiu (0 o 1)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Resultats per pàgina",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Llista de proveïdors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Provider"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Providers"
                ],
                "summary": "Crear un nou proveïdor",
                "operationId": "8d61f862240fcb54ca4379f8211c756f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Provider"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveïdor creat"
                    },
                    "422": {
                        "description": "Error de validació"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/providers/{id}": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Detalls d'un proveïdor",
                "operationId": "2e8a9bc6272603d6fc877d77d9472c78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dades del proveïdor"
                    },
                    "404": {
                        "description": "Proveïdor no trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Providers"
                ],
                "summary": "Actualitzar un proveïdor existent",
                "operationId": "9f26750a48584c4e9d4a861e4d024d1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nou Nom Proveïdor"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualitzat correctament"
                    },
                    "404": {
                        "description": "Proveïdor no trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Providers"
                ],
                "summary": "Esborrar un proveïdor",
                "operationId": "1aa03e6dae527bbdd64b73a803e0a142",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Esborrat correctament"
                    },
                    "404": {
                        "description": "Proveïdor no trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/password/forgot": {
            "post": {
                "tags": [
                    "Recovery"
                ],
                "summary": "Sol·licitar enllaç de recuperació de contrasenya",
                "operationId": "d67c4a8e7fd634ef49cf68e6cf4e1b56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuari@exemple.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enllaç enviat (o simulat per seguretat)"
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Recovery"
                ],
                "summary": "Restablir contrasenya amb token",
                "operationId": "853d536361840cd9a46f301426cf48a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contrasenya restablerta"
                    },
                    "400": {
                        "description": "Token o dades invàlides"
                    }
                }
            }
        },
        "/api/email/resend-verification": {
            "post": {
                "tags": [
                    "Recovery"
                ],
                "summary": "Reenviar email de verificació",
                "operationId": "9d413259bbc1f5e08544fe11c8f136fc",
                "responses": {
                    "200": {
                        "description": "Email enviat"
                    },
                    "401": {
                        "description": "No autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Provider": {
                "title": "Provider",
                "description": "Provider model definition",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Tech Solutions S.L."
                    },
                    "contact_person": {
                        "type": "string",
                        "example": "Joan Garcia",
                        "nullable": true
                    },
                    "contact_email": {
                        "type": "string",
                        "format": "email",
                        "example": "joan@techsolutions.com",
                        "nullable": true
                    },
                    "contact_phone": {
                        "type": "string",
                        "example": "938887766",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "Carrer Major 15, Barcelona",
                        "nullable": true
                    },
                    "tax_id": {
                        "type": "string",
                        "example": "B12345678",
                        "nullable": true
                    },
                    "website": {
                        "type": "string",
                        "format": "url",
                        "example": "https://techsolutions.com",
                        "nullable": true
                    },
                    "provider_type": {
                        "description": "rental, lifetime, etc.",
                        "type": "string",
                        "example": "rental"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}